backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
django-filter
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv